requirements.txt
Flask==0.10.1
pep8==1.7.0
pylint==1.5.4
pytest==2.8.7
simplecoremidi==0.3
